+++ CHANGES.md
@@ -1,5 +1,6 @@
 # 6.0.41 2022-xx-xx
  - 2022-03-31 Fixed check of UntilTime in Kernel::System::Ticket::Event::TicketPendingTimeReset. [#221)(https://github.com/znuny/Znuny/issues/221)
+ - 2022-03-23 Removed configurable system commands from generic agents (CVE-2021-36100).
  - 2022-03-14 Fixed sending notifications to invalid customer users.
 
 # 6.0.40 2022-03-09